Fix rules error handings
